Verificarea parolei in portal/admin:materii
Plus doua modificari minore

+++ portal/subpagini/admin/materii.post.php
@@ -1,8 +1,8 @@
 <?php
 
-var_dump($_SESSION);
-
 $db = new db_connection();
+
+$response = new stdClass();
 
 if (isset($_POST["form-id"])) {
 
@@ -13,16 +13,44 @@ if (isset($_POST["form-id"])) {
 
 			$db->insert_materie(array("Nume" => $_POST["nume"]));
 
+			$response->status = "success";
+
 		}
 
 		if (isset($_POST["sterge-materie"])) {
 
-			$db->delete_materie($_POST["materie-id"]);
+			// verifica parola
+
+			$user = $db->retrieve_utilizator_where_id("Id,Parola", $_POST["admin-id"]);
+			$passwordCorrect = password_verify($_POST["password"], $user["Parola"]);
+
+			if ($passwordCorrect) {
+
+				$db->delete_materie($_POST["materie-id"]);
+
+				$response->status = "success";
+
+			} else {
+
+				$response->status = "password-failed";
+
+			}
 
 		}
 
+	} else {
+
+		$response->status = "form-id-failed";
+
 	}
 
+} else {
+
+	$response->status = "form-id-not-found";
+
 }
+
+header("Content-type: application/json");
+echo(json_encode($response));
 
 ?>
